# openapi.yaml
openapi: 3.1.0
info:
  title: Lume CRM API
  version: 1.0.0
  description: |
    API specification for Lume CRM - a customer relationship management system for window film sales.
    
    ## Authentication
    
    The API uses cookie-based authentication with JWT tokens:
    - `crm-token` - Short-lived access token (15 min)
    - `crm-refresh-token` - Long-lived refresh token (7 days)
    
    Alternatively, Bearer tokens can be used via the `Authorization` header.
    
    ## Rate Limiting
    
    Login endpoint is rate-limited to 10 requests per 15 minutes per IP.
    
    ## Error Responses
    
    All endpoints return consistent error responses:
    ```json
    {
      "error": "string",
      "code": "string",
      "details": "string",
      "hint": "string"
    }
    ```
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://lume-crm.vercel.app
    description: Production server
security:
  - cookieAuth: []
  - bearerAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: crm-token
      description: JWT access token stored in httpOnly cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token in Authorization header
  schemas:
...
    MeResponse:
      type: object
      properties:
        authenticated:
          type: boolean
        email:
          type: string
      example:
        authenticated: true
        email: "vendedor@lume.local"
    MeErrorResponse:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Token nao fornecido"
...
paths:
...
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Returns authenticated user info from JWT token (cookie or Bearer header)
      operationId: getCurrentUser
      responses:
        '200':
          description: User authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeErrorResponse'
...